+++ src/modules/Elsa.JavaScript/Handlers/ConfigureEngineWithCommonFunctions.cs
@@ -1,0 +1,69 @@
+using System.Text.Encodings.Web;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using System.Text.RegularExpressions;
+using System.Text.Unicode;
+using Elsa.Extensions;
+using Elsa.JavaScript.Notifications;
+using Elsa.Mediator.Contracts;
+using JetBrains.Annotations;
+
+namespace Elsa.JavaScript.Handlers;
+
+/// <summary>
+/// 
+/// </summary>
+[UsedImplicitly] 
+public class ConfigureEngineWithCommonFunctions : INotificationHandler<EvaluatingJavaScript>
+{
+    private readonly JsonSerializerOptions _jsonSerializerOptions = CreateJsonSerializerOptions();
+
+    /// <inheritdoc />
+    public Task HandleAsync(EvaluatingJavaScript notification, CancellationToken cancellationToken)
+    {
+        var engine = notification.Engine;
+        var context = notification.Context;
+        
+        // Add common functions.
+        engine.SetValue("getWorkflowDefinitionId", (Func<string>)(() => context.GetWorkflowExecutionContext().Workflow.Identity.DefinitionId));
+        engine.SetValue("getWorkflowDefinitionVersionId", (Func<string>)(() => context.GetWorkflowExecutionContext().Workflow.Identity.Id));
+        engine.SetValue("getWorkflowDefinitionVersion", (Func<int>)(() => context.GetWorkflowExecutionContext().Workflow.Identity.Version));
+        engine.SetValue("getWorkflowInstanceId", (Func<string>)(() => context.GetActivityExecutionContext().WorkflowExecutionContext.Id));
+        engine.SetValue("setCorrelationId", (Action<string?>)(value => context.GetActivityExecutionContext().WorkflowExecutionContext.CorrelationId = value));
+        engine.SetValue("getCorrelationId", (Func<string?>)(() => context.GetActivityExecutionContext().WorkflowExecutionContext.CorrelationId));
+        engine.SetValue("setVariable", (Action<string, object>)((name, value) => context.SetVariableInScope(name, value)));
+        engine.SetValue("getVariable", (Func<string, object?>)(name => context.GetVariableInScope(name)));
+        engine.SetValue("getInput", (Func<string, object?>)(name => context.GetInput(name)));
+        engine.SetValue("getOutputFrom", (Func<string, string?, object?>)((activityIdName, outputName) => context.GetOutput(activityIdName, outputName)));
+        engine.SetValue("getLastResult", (Func<object?>)(() => context.GetLastResult()));
+        engine.SetValue("isNullOrWhiteSpace", (Func<string, bool>)(value => string.IsNullOrWhiteSpace(value)));
+        engine.SetValue("isNullOrEmpty", (Func<string, bool>)(value => string.IsNullOrEmpty(value)));
+        engine.SetValue("toJson", (Func<object, string>)(value => Serialize(value)));
+        engine.SetValue("parseGuid", (Func<string, Guid>)(value => Guid.Parse(value)));
+        engine.SetValue("newGuid", (Func<Guid>)(() => Guid.NewGuid()));
+        engine.SetValue("newGuidString", (Func<string>)(() => Guid.NewGuid().ToString()));
+        engine.SetValue("newShortGuid", (Func<string>)(() => Regex.Replace(Convert.ToBase64String(Guid.NewGuid().ToByteArray()), "[/+=]", "")));
+        
+        // Deprecated, use newGuidString instead.
+        engine.SetValue("getGuidString", (Func<string>)(() => Guid.NewGuid().ToString()));
+
+        // Deprecated, use newShortGuid instead.
+        engine.SetValue("getShortGuid", (Func<string>)(() => Regex.Replace(Convert.ToBase64String(Guid.NewGuid().ToByteArray()), "[/+=]", "")));
+        return Task.CompletedTask;
+    }
+    
+    private string Serialize(object value)
+    {
+        return JsonSerializer.Serialize(value, _jsonSerializerOptions);
+    }
+    
+    private static JsonSerializerOptions CreateJsonSerializerOptions()
+    {
+        var options = new JsonSerializerOptions
+        {
+            Encoder = JavaScriptEncoder.Create(UnicodeRanges.All)
+        };
+        options.Converters.Add(new JsonStringEnumConverter());
+        return options;
+    }
+}
