+++ src/modules/Elsa.JavaScript/Handlers/ConfigureEngineWithCommonTypes.cs
@@ -1,0 +1,27 @@
+using Elsa.Extensions;
+using Elsa.JavaScript.Notifications;
+using Elsa.Mediator.Contracts;
+using JetBrains.Annotations;
+
+namespace Elsa.JavaScript.Handlers;
+
+/// <summary>
+/// 
+/// </summary>
+[UsedImplicitly]
+public class ConfigureEngineWithCommonTypes : INotificationHandler<EvaluatingJavaScript>
+{
+    /// <inheritdoc />
+    public Task HandleAsync(EvaluatingJavaScript notification, CancellationToken cancellationToken)
+    {
+        var engine = notification.Engine;
+        
+        // Add common .NET types.
+        engine.RegisterType<DateTime>();
+        engine.RegisterType<DateTimeOffset>();
+        engine.RegisterType<TimeSpan>();
+        engine.RegisterType<Guid>();
+        
+        return Task.CompletedTask;
+    }
+}
