+++ src/modules/Elsa.JavaScript/Handlers/ConfigureEngineWithVariableAccessors.cs
@@ -1,0 +1,31 @@
+using Elsa.Extensions;
+using Elsa.JavaScript.Notifications;
+using Elsa.Mediator.Contracts;
+using Humanizer;
+using JetBrains.Annotations;
+
+namespace Elsa.JavaScript.Handlers;
+
+/// <summary>
+/// 
+/// </summary>
+[UsedImplicitly]
+public class ConfigureEngineWithVariableAccessors : INotificationHandler<EvaluatingJavaScript>
+{
+    /// <inheritdoc />
+    public Task HandleAsync(EvaluatingJavaScript notification, CancellationToken cancellationToken)
+    {
+        var engine = notification.Engine;
+        var context = notification.Context;
+        var variableNames = context.GetVariableNamesInScope().ToList();
+
+        foreach (var variableName in variableNames)
+        {
+            var pascalName = variableName.Pascalize();
+            engine.SetValue($"get{pascalName}", (Func<object?>)(() => context.GetVariableInScope(variableName)));
+            engine.SetValue($"set{pascalName}", (Action<object?>)(value => context.SetVariableInScope(variableName, value)));
+        }
+        
+        return Task.CompletedTask;
+    }
+}
