+++ src/modules/Elsa.JavaScript/Providers/JavaScriptExpressionSyntaxProvider.cs
@@ -2,10 +2,12 @@ using Elsa.Expressions.Contracts;
 using Elsa.Expressions.Models;
 using Elsa.Extensions;
 using Elsa.JavaScript.Expressions;
+using JetBrains.Annotations;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Elsa.JavaScript.Providers;
 
+[UsedImplicitly]
 internal class JavaScriptExpressionDescriptorProvider : IExpressionDescriptorProvider
 {
     private const string TypeName = "JavaScript";
