+++ src/modules/Elsa.JavaScript/Services/JintJavaScriptEvaluator.cs
@@ -1,24 +1,15 @@
-using System.Diagnostics.CodeAnalysis;
 using System.Security.Cryptography;
 using System.Text;
-using System.Text.Encodings.Web;
-using System.Text.Json;
-using System.Text.Json.Serialization;
-using System.Text.RegularExpressions;
-using System.Text.Unicode;
 using Elsa.Expressions.Helpers;
 using Elsa.Expressions.Models;
-using Elsa.Extensions;
 using Elsa.JavaScript.Contracts;
 using Elsa.JavaScript.Helpers;
 using Elsa.JavaScript.Notifications;
 using Elsa.JavaScript.ObjectConverters;
 using Elsa.JavaScript.Options;
 using Elsa.Mediator.Contracts;
-using Humanizer;
+using Esprima.Ast;
 using Jint;
-using Jint.Native;
-using Jint.Native.TypedArray;
 using Jint.Runtime.Interop;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Configuration;
@@ -34,7 +25,6 @@ public class JintJavaScriptEvaluator(IConfiguration configuration, INotification
     : IJavaScriptEvaluator
 {
     private readonly JintOptions _jintOptions = scriptOptions.Value;
-    private readonly JsonSerializerOptions _jsonSerializerOptions = CreateJsonSerializerOptions();
 
     /// <inheritdoc />
     public async Task<object?> EvaluateAsync(string expression,
@@ -56,164 +46,86 @@ public class JintJavaScriptEvaluator(IConfiguration configuration, INotification
 
         var engine = new Engine(opts =>
         {
-            if (_jintOptions.AllowClrAccess)
-                opts.AllowClr();
-
-            // Wrap objects in ObjectWrapper instances and set their prototype to Array.prototype if they are array-like.
-            opts.SetWrapObjectHandler((engine, target, type) =>
-            {
-                var instance = ObjectWrapper.Create(engine, target);
-
-                if (ObjectArrayHelper.DetermineIfObjectIsArrayLikeClrCollection(target.GetType()))
-                    instance.Prototype = engine.Intrinsics.Array.PrototypeObject;
-
-                return instance;
-            });
-            
-            opts.Interop.ObjectConverters.AddRange([new ByteArrayConverter(), new ExpandoObjectConverter()]);
+            ConfigureClrAccess(opts);
+            ConfigureObjectWrapper(opts);
+            ConfigureObjectConverters(opts);
         });
 
         configureEngine?.Invoke(engine);
-
-        // Add common functions.
-        engine.SetValue("getWorkflowDefinitionId", (Func<string>)(() => context.GetWorkflowExecutionContext().Workflow.Identity.DefinitionId));
-        engine.SetValue("getWorkflowDefinitionVersionId", (Func<string>)(() => context.GetWorkflowExecutionContext().Workflow.Identity.Id));
-        engine.SetValue("getWorkflowDefinitionVersion", (Func<int>)(() => context.GetWorkflowExecutionContext().Workflow.Identity.Version));
-        engine.SetValue("getWorkflowInstanceId", (Func<string>)(() => context.GetActivityExecutionContext().WorkflowExecutionContext.Id));
-        engine.SetValue("setCorrelationId", (Action<string?>)(value => context.GetActivityExecutionContext().WorkflowExecutionContext.CorrelationId = value));
-        engine.SetValue("getCorrelationId", (Func<string?>)(() => context.GetActivityExecutionContext().WorkflowExecutionContext.CorrelationId));
-        engine.SetValue("setVariable", (Action<string, object>)((name, value) => context.SetVariableInScope(name, value)));
-        engine.SetValue("getVariable", (Func<string, object?>)(name => context.GetVariableInScope(name)));
-        engine.SetValue("getInput", (Func<string, object?>)(name => context.GetInput(name)));
-        engine.SetValue("getOutputFrom", (Func<string, string?, object?>)((activityIdName, outputName) => context.GetOutput(activityIdName, outputName)));
-        engine.SetValue("getLastResult", (Func<object?>)(() => context.GetLastResult()));
-
-        // Create variable getters and setters for each variable.
-        CreateVariableAccessors(engine, context);
-
-        // Create workflow input accessors - only if the current activity is not part of a composite activity definition.
-        // Otherwise, the workflow input accessors will hide the composite activity input accessors which rely on variable accessors created above.
-        CreateWorkflowInputAccessors(engine, context);
-
-        // Create output getters for each activity.
-        await CreateActivityOutputAccessorsAsync(engine, context);
-
-        // Create argument getters for each argument.
-        foreach (var argument in options.Arguments)
-            engine.SetValue($"get{argument.Key}", (Func<object?>)(() => argument.Value));
-
-        // Add common functions.
-        engine.SetValue("isNullOrWhiteSpace", (Func<string, bool>)(value => string.IsNullOrWhiteSpace(value)));
-        engine.SetValue("isNullOrEmpty", (Func<string, bool>)(value => string.IsNullOrEmpty(value)));
-        engine.SetValue("toJson", (Func<object, string>)(value => Serialize(value)));
-        engine.SetValue("parseGuid", (Func<string, Guid>)(value => Guid.Parse(value)));
-        engine.SetValue("newGuid", (Func<Guid>)(() => Guid.NewGuid()));
-        engine.SetValue("newGuidString", (Func<string>)(() => Guid.NewGuid().ToString()));
-        engine.SetValue("newShortGuid", (Func<string>)(() => Regex.Replace(Convert.ToBase64String(Guid.NewGuid().ToByteArray()), "[/+=]", "")));
-
-        // Deprecated, use newGuidString instead.
-        engine.SetValue("getGuidString", (Func<string>)(() => Guid.NewGuid().ToString()));
-
-        // Deprecated, use newShortGuid instead.
-        engine.SetValue("getShortGuid", (Func<string>)(() => Regex.Replace(Convert.ToBase64String(Guid.NewGuid().ToByteArray()), "[/+=]", "")));
-
-        // Create configuration value accessor
-        if (_jintOptions.AllowConfigurationAccess)
-            engine.SetValue("getConfig", (Func<string, object?>)(name => configuration.GetSection(name).Value));
-
-        // Add common .NET types.
-        engine.RegisterType<DateTime>();
-        engine.RegisterType<DateTimeOffset>();
-        engine.RegisterType<TimeSpan>();
-        engine.RegisterType<Guid>();
-
-        // Invoke registered configuration callback.
+        ConfigureArgumentGetters(engine, options);
+        ConfigureConfigurationAccess(engine);
         _jintOptions.ConfigureEngineCallback(engine, context);
-
-        // Allow listeners invoked by the mediator to configure the engine.
         await mediator.SendAsync(new EvaluatingJavaScript(engine, context), cancellationToken);
 
         return engine;
     }
 
-    private void CreateWorkflowInputAccessors(Engine engine, ExpressionExecutionContext context)
+    private void ConfigureClrAccess(Jint.Options options)
     {
-        if (context.IsInsideCompositeActivity())
-            return;
-
-        var inputs = context.GetWorkflowInputs().ToDictionary(x => x.Name);
-
-        if (!context.TryGetWorkflowExecutionContext(out var workflowExecutionContext))
-            return;
-
-        var inputDefinitions = workflowExecutionContext.Workflow.Inputs;
-
-        foreach (var inputDefinition in inputDefinitions)
+        if (_jintOptions.AllowClrAccess)
+            options.AllowClr();
+    }
+    
+    private void ConfigureObjectWrapper(Jint.Options options)
+    {
+        options.SetWrapObjectHandler((engine, target, type) =>
         {
-            var input = inputs.GetValueOrDefault(inputDefinition.Name);
-            engine.SetValue($"get{inputDefinition.Name}", (Func<object?>)(() => input?.Value));
-        }
+            var instance = ObjectWrapper.Create(engine, target);
+
+            if (ObjectArrayHelper.DetermineIfObjectIsArrayLikeClrCollection(target.GetType()))
+                instance.Prototype = engine.Intrinsics.Array.PrototypeObject;
+
+            return instance;
+        });
     }
 
-    [RequiresUnreferencedCode("Calls Jint.Engine.SetValue<T>(String, T)")]
-    private static void CreateVariableAccessors(Engine engine, ExpressionExecutionContext context)
+    private void ConfigureObjectConverters(Jint.Options options)
     {
-        var variableNames = context.GetVariableNamesInScope().ToList();
-
-        foreach (var variableName in variableNames)
-        {
-            var pascalName = variableName.Pascalize();
-            engine.SetValue($"get{pascalName}", (Func<object?>)(() => context.GetVariableInScope(variableName)));
-            engine.SetValue($"set{pascalName}", (Action<object?>)(value => context.SetVariableInScope(variableName, value)));
-        }
+        options.Interop.ObjectConverters.AddRange([new ByteArrayConverter(), new ExpandoObjectConverter()]);
     }
 
-    private static async Task CreateActivityOutputAccessorsAsync(Engine engine, ExpressionExecutionContext context)
+    private void ConfigureArgumentGetters(Engine engine, ExpressionEvaluatorOptions options)
     {
-        var activityOutputs = context.GetActivityOutputs();
-
-        await foreach (var activityOutput in activityOutputs)
-        foreach (var outputName in activityOutput.OutputNames)
-            engine.SetValue($"get{outputName}From{activityOutput.ActivityName}", (Func<object?>)(() => context.GetOutput(activityOutput.ActivityId, outputName)));
+        foreach (var argument in options.Arguments)
+            engine.SetValue($"get{argument.Key}", (Func<object?>)(() => argument.Value));
+    }
+    
+    private void ConfigureConfigurationAccess(Engine engine)
+    {
+        if (_jintOptions.AllowConfigurationAccess)
+            engine.SetValue("getConfig", (Func<string, object?>)(name => configuration.GetSection(name).Value));
     }
 
     private object? ExecuteExpressionAndGetResult(Engine engine, string expression)
     {
+        var preparedScript = GetOrCreatePrepareScript(expression);
+        var result = engine.Evaluate(preparedScript);
+        return result.ToObject();
+    }
+
+    private Prepared<Script> GetOrCreatePrepareScript(string expression)
+    {
         var cacheKey = "jint:script:" + Hash(expression);
 
-        var parsedScript = memoryCache.GetOrCreate(cacheKey, entry =>
+        return memoryCache.GetOrCreate(cacheKey, entry =>
         {
             if (_jintOptions.ScriptCacheTimeout.HasValue)
                 entry.SetAbsoluteExpiration(_jintOptions.ScriptCacheTimeout.Value);
 
-            var prepareOptions = new ScriptPreparationOptions
-            {
-                ParsingOptions = new ScriptParsingOptions
-                {
-                    AllowReturnOutsideFunction = true
-                }
-            };
-            return Engine.PrepareScript(expression, options: prepareOptions);
+            return PrepareScript(expression);
         })!;
-
-        var result = engine.Evaluate(parsedScript);
-        return result.ToObject();
     }
 
-    [RequiresUnreferencedCode("Calls System.Text.Json.JsonSerializer.Serialize<TValue>(TValue, JsonSerializerOptions)")]
-    private string Serialize(object value)
+    private Prepared<Script> PrepareScript(string expression)
     {
-        return JsonSerializer.Serialize(value, _jsonSerializerOptions);
-    }
-    
-    private static JsonSerializerOptions CreateJsonSerializerOptions()
-    {
-        var options = new JsonSerializerOptions
+        var prepareOptions = new ScriptPreparationOptions
         {
-            Encoder = JavaScriptEncoder.Create(UnicodeRanges.All)
+            ParsingOptions = new ScriptParsingOptions
+            {
+                AllowReturnOutsideFunction = true
+            }
         };
-        options.Converters.Add(new JsonStringEnumConverter());
-        return options;
+        return Engine.PrepareScript(expression, options: prepareOptions);
     }
     
     private string Hash(string input)
